mental-health-diagnostics/requirements.txt
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn
pydantic
python-dotenv
